% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french.R
\docType{data}
\name{french}
\alias{french}
\title{French Elites Social Networks and Multivariate Attributes}
\format{
List including a binary adjacency matrix and a binary mutlivariate attributes
}
\usage{
french
}
\description{
The dataset contains a social network of french financial elites and their multivariate attributes
It includes social interaction between 28 elites and their binary responses to 13 questions.
The data were downloaded from the social network Repository created by Prof. Linton Freeman.
}
\details{
social network and multivariate attributes
}
\keyword{datasets}
