% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gofblsm.R
\name{Gofblsm}
\alias{Gofblsm}
\title{Assess the fit of the BLSM}
\usage{
Gofblsm(model, Y.ia)
}
\arguments{
\item{model}{object of class BLSM}

\item{Y.ia}{N by M matrix containing the binary item response matrix}
}
\value{
scalar containing:
 \itemize{
 \item \code{Ya.auc} scaler of the area under the curve for the multivariate covariates
 }
}
\description{
assess the fit of the model using ROC curves and auc values
}
\examples{
attach(french)
a=blsm(Niter=5,Y.ia,D=2)
Gofblsm(a,Y.ia)
}
