% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulateaplsm.R
\name{Simulateaplsm}
\alias{Simulateaplsm}
\title{Simulate from the APLSM}
\usage{
Simulateaplsm(model, type)
}
\arguments{
\item{model}{object of class APlsm}

\item{type}{character indicating the types of model. It could be "DD", distance by distance model, "DV", distance by vector model,
"VV", vector by vector model}
}
\value{
list containing:
 \itemize{
 \item \code{Y.i} (\code{N} x \code{N}) matrix containing the simulated Y.i
 \item \code{Y.ia} (\code{N} x \code{M}) matrix containing the simulated Y.ia
 }
}
\description{
function to simulate networks from the APLSM
}
\examples{
attach(french)
b=aplsm(Niter=3,Y.i, Y.ia,D=2, type="DD")
Simulateaplsm(b,"DD")
}
