% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulateblsm.R
\name{Simulateblsm}
\alias{Simulateblsm}
\title{Simulate from the BLSM model}
\usage{
Simulateblsm(model)
}
\arguments{
\item{model}{object of class BLSM}
}
\value{
list containing:
 \itemize{
 \item \code{Y.ia} (\code{N} x \code{M}) matrix containing the simulated Y.ia
 }
}
\description{
function to simulate networks from the BLSM
}
\examples{
attach(french)
a=blsm(Niter=5,Y.ia,D=2)
Simulateblsm(a)
}
